updated CLCore + Added FunctionMeasure
+ Fixed Sum sample

+++ src/Sum/Sum.h
@@ -1,3 +1,3 @@
 #pragma once
 
-cl_int sum( cl_int *input, int size, ICLProgram* pProgram );
+cl_int sum( cl_int* input, size_t size, ICLProgram* pProgram );
